SPU2-X: Some bugfixes to reverb that were introduced in my last commit.  Still some minor problems in this build.  I'll fix them soon.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@527 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/spu2-x/src/Win32/SndOut_DSound.cpp
@@ -144,9 +144,9 @@ public:
 		memset(&wfx, 0, sizeof(WAVEFORMATEX)); 
 		wfx.wFormatTag		= WAVE_FORMAT_PCM;
 		wfx.nSamplesPerSec	= SampleRate;
-		wfx.nChannels		= speakerConfig;
+		wfx.nChannels		= (WORD)speakerConfig;
 		wfx.wBitsPerSample	= 16;
-		wfx.nBlockAlign		= 2*speakerConfig;
+		wfx.nBlockAlign		= 2*(WORD)speakerConfig;
 		wfx.nAvgBytesPerSec	= SampleRate * wfx.nBlockAlign;
 		wfx.cbSize			= 0;
